permuted

+++ Segmentation/segment.py
@@ -109,7 +109,7 @@ class PoseDataset(torch.utils.data.Dataset):
 
 def tensor_to_image(tensor):
     tensor = tensor*255
-    tensor = np.array(tensor, dtype=np.uint8)
+    tensor = np.array(tensor)
     if np.ndim(tensor)>3:
         assert tensor.shape[0] == 1
         tensor = tensor[0]
